ml-pipeline/data_generation/requirements-minimal.txt
# Minimal requirements for data generation
openai>=1.3.0
rich>=13.0.0
loguru>=0.7.0
python-dotenv>=1.0.0